Moved Moduletabs and ExpandButton component out of librarymodule. Fixed imports
