+++ src/client/src/modules/Inspector_V2/header/InspectorModuleHeader_V2.styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import {Color} from "../../../assets/color/Color";
+import {Color} from "assets/color/Color";
 
 
 export const InspectorModuleHeader_V2Styled = styled.div`
