+++ src/client/src/modules/Inspector_V2/header/InspectorModuleHeader_V2.tsx
@@ -1,6 +1,6 @@
-import {InspectorTab} from "../../../models/enums/InspectorTab";
+import {InspectorTab} from "models";
 import {InspectorModuleHeader_V2Styled} from "./InspectorModuleHeader_V2.styled";
-import {ModuleTabs} from "../../library/components/header/components/ModuleTabs";
+import {ModuleTabs} from "components/menus/tabMenu/ModuleTabs";
 
 interface Props {
     id: string;
@@ -12,6 +12,6 @@ interface Props {
 
 export const InspectorModuleHeader_V2 = ({id, isModuleOpen, activeTab, setActiveTab, setModuleOpenStatus}: Props) => (
     <InspectorModuleHeader_V2Styled>
-        <ModuleTabs id={id} activeTab={activeTab} setActiveTab={setActiveTab} onOpen={setModuleOpenStatus} />
+        <ModuleTabs id={id} activeTab={activeTab} setActiveTab={setActiveTab} onOpen={setModuleOpenStatus}/>
     </InspectorModuleHeader_V2Styled>
 );
