+++ src/client/src/modules/Inspector_V2/header/components/InspectorModuleTabSection_V2.tsx
@@ -1,13 +1,14 @@
 import {Inspector_V2TabsWrapper} from "./InspectorModuleTabSection_V2.styled";
-import {InspectorTab} from "../../../../models/enums/InspectorTab";
+import {InspectorTab} from "models/enums/InspectorTab";
 
 interface Props {
     activeTab: InspectorTab,
     setActiveTab: (tab: InspectorTab) => void,
     onOpen: () => void;
 }
-export const InspectorModuleTabSection_V2 = ({activeTab, setActiveTab, onOpen}: Props) => (
-  <Inspector_V2TabsWrapper>
 
-  </Inspector_V2TabsWrapper>
+export const InspectorModuleTabSection_V2 = ({activeTab, setActiveTab, onOpen}: Props) => (
+    <Inspector_V2TabsWrapper>
+
+    </Inspector_V2TabsWrapper>
 );
